:bug: 修复了用户领取优惠券的bug

+++ src/main/java/com/o0u0o/missyou/model/Coupon.java
@@ -28,10 +28,10 @@ public class Coupon extends BaseEntity {
 
     /** 优惠券主键ID */
     @Id
-    private Integer id;
+    private Long id;
 
     /** 活动ID */
-    private Integer activityId;
+    private Long activityId;
 
     /** 优惠券标题 */
     private String title;
